Cnfig loading DRY, refactor
